Allow debug builds to coexist with releases
The debug build has a different id (com.affectiva.affdexme.debug) so
android treats it as a completely different app.  By default, though,
both apps look the same on the app launch screen so the debug version is
now called AffdexMeDbg so you can have both at the same time and be able
to tell which is which.