Bump versionCode higher than the last release
We evidently had some build configuration that set the versionCode to
the jenkins build number which at the point we switched to the new
AffdexMe process was 491.  I think we want to manage the versionCode
manually from now on, so I'll set it one higher than what it was and we
can increment it when we think it's important to do so.