Make apk filename more descriptive
It was just "app-debug.apk" or "app-release.apk".  This commit changes
the "app" to "AffdexMe" and adds the human-readable versionName and
integer versionCode to the filename.

I would have liked the filename to track the cosmetic name (i.e., have
the debug file be called "AffdexMeDbg-....apk") but that turned out to
be surprisingly difficult given the way the build system is set up.  I
think it should be possible but this is the best bang for the buck for
the amount of effort I'm willing to put in at this point.